+++ spec/support/kitchen/terraform/config_schemas/boolean_examples.rb
@@ -14,20 +14,42 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "support/kitchen/terraform/config_attribute_context"
-
 ::RSpec.shared_examples "Kitchen::Terraform::ConfigSchemas::Boolean" do |attribute:|
-  include_context "Kitchen::Terraform::ConfigAttribute", attribute: attribute do
-    context "when the config omits #{attribute.inspect}" do
-      it_behaves_like "a default value is used", default_value: true
+  context "when the config omits #{attribute.inspect}" do
+    subject do
+      described_class.new
     end
 
-    context "when the config associates #{attribute.inspect} with a nonboolean" do
-      it_behaves_like "the value is invalid", error_message: /#{attribute}.*must be boolean/, value: "abc"
+    before do
+      described_class.validations.fetch(attribute).call attribute, subject[attribute], subject
     end
 
-    context "when the config associates #{attribute.inspect} with a boolean" do
-      it_behaves_like "the value is valid", value: false
+    specify "should associate #{attribute.inspect} with true" do
+      expect(subject[attribute]).to be true
+    end
+  end
+
+  context "when the config associates #{attribute.inspect} with a nonboolean" do
+    subject do
+      described_class.new attribute => "abc"
+    end
+
+    specify "should raise a Kitchen::UserError" do
+      expect do
+        described_class.validations.fetch(attribute).call attribute, subject[attribute], subject
+      end.to raise_error ::Kitchen::UserError, /#{attribute}.*must be boolean/
+    end
+  end
+
+  context "when the config associates #{attribute.inspect} with a boolean" do
+    subject do
+      described_class.new attribute => false
+    end
+
+    specify "should not raise a Kitchen::UserError" do
+      expect do
+        described_class.validations.fetch(attribute).call attribute, subject[attribute], subject
+      end.not_to raise_error
     end
   end
 end